comm_size now returns size_t type to prevent any possible overflow when including comm_size in expressions at high process count

+++ include/comm_quda.h
@@ -168,7 +168,7 @@ extern "C" {
   /**
      @return Number of processes
   */
-  int comm_size(void);
+  size_t comm_size(void);
 
   /**
      @return GPU id associated with this process
